refactor: 패키지 이동 및 사용 하지 않는 코드 제거
