+++ app/src/main/java/woowacourse/shopping/domain/model/CartItem.kt
@@ -4,11 +4,3 @@ data class CartItem(val id: Long, val product: Product, val quantity: Int = 1) {
     val price: Int
         get() = product.price * quantity
 }
-
-fun CartItem.incrementQuantity(incrementBy: Int): CartItem {
-    return this.copy(quantity = this.quantity + incrementBy)
-}
-
-fun CartItem.decrementQuantity(decrementBy: Int): CartItem {
-    return this.copy(quantity = this.quantity - decrementBy)
-}
