+++ app/src/main/java/woowacourse/shopping/utils/Event.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.view
+package woowacourse.shopping.utils
 
 open class Event<out T>(private val content: T) {
     var hasBeenHandled = false
